requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
diffusers==0.29.0
docker-pycreds==0.4.0
filelock==3.13.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.2.0
platformdirs==4.2.2
protobuf==5.27.1
psutil==6.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentry-sdk==2.6.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.2
zipp==3.19.2
